requirements.txt
panda3d==1.10.6
h5py==2.10.0
Pillow==7.1.2
tqdm==4.46.1
numpy==1.19.0
pygame==1.9.6